refactor: Refactor api and lib directory
