+++ src/app/api/getWords/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { prisma } from '../../lib/client';
+import { prisma } from '../../../lib/client';
 
 export const dynamic = 'force-dynamic';
