+++ src/app/api/updateWord/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { prisma } from '../../lib/client';
+import { prisma } from '../../../lib/client';
 
 export const dynamic = 'force-dynamic';
 
@@ -7,9 +7,15 @@ export async function POST(req: NextRequest) {
   try {
     const requestBody = await req.json();
     const { id } = requestBody;
-    const newEntry = await prisma.dictionary.delete({
+    const { wordName } = requestBody;
+    const { meaning } = requestBody;
+    const newEntry = await prisma.dictionary.update({
       where: {
         id,
+      },
+      data: {
+        wordName,
+        meaning,
       },
     });
