+++ src/app/lib/interface.ts
